refactor: split generator tasks in separated files for readability and maintainability
