+++ cli/templates/base/app/main.tsx
@@ -1,22 +1,1 @@
-import React, { useEffect } from 'react';
-import { createRoot } from 'react-dom/client';
-
-// Root component.
-function Root() {
-  useEffect(() => {
-    // Hide the welcome banner.
-    const banner = document.querySelector('#welcome-banner')!;
-    banner.classList.add('dismissed');
-    setTimeout(() => banner.remove(), 500);
-  }, []);
-
-  return (
-    <div>
-      <p>Hello World</p>
-    </div>
-  );
-}
-
-const rootNode = document.querySelector('#app-container')!;
-const root = createRoot(rootNode);
-root.render(<Root />);
+// This entry point serves as a placeholder and will be substituted with the component library's variant entry points.
